OMS Core Elixir v0.0.1 Omscore.Auth

The Auth context. This accumulates functionalities around logging in, the user datastructure and everything password related

Functions

authenticate_user(username, plain_text_password)

Check if a user exists in the database and if that user has the password and is activated.

Returns {:ok, user}

change_user(user)

Creates a user Ecto.Changeset and returns it

check_access_token(access_token)

Check an access token for validity

Just requires a check with the Guardian library by cryptographic means, no db access is performed.

Returns {:ok, user, claims} or {:error, any}

check_refresh_token(refresh_token)

Check a refresh token for validity

First checks if the token is a valid token by cryptographic means and then checks if it is in the db (if not the user was logged out).

Returns {:ok, user, token_db} or {:error, any}

check_user_existence(username)

Checks if a given username exists in the database

Returns true or false

create_access_token(user, refresh_token_db)

Creates an access token

Access tokens are shortlived and not saved in db, they are only cryptographically signed.

Returns {:ok, access_token, claims}, but you can ignore the latter.

create_password_reset_object(user)

Creates a password reset object.

Since the url is stored in hashed form, it also returns the original url, which can then be sent to the user. Please don't store this url anywhere, as it could be used by an attacker to gain full system access.

Returns {:ok, password_reset, url}

create_refresh_token(user, device)

Creates a refresh token

Create a longlived refresh token and saves it to db This can be used to create access tokens later on.

Returns {:ok, refresh_token, refresh_token_db}. The refresh_token is the actual token and refresh_token_db is an instance of Omscore.Auth.RefreshToken

create_user(attrs \\ %{})

Creates a new user by attributes.

See Omscore.Auth.User for attrs.

Returns {:ok, user} with the new user.

delete_user(user)

Deletes a user.

This will cascade up and also delete the associated member, and from there to its memberships, join requests and other associated data structures. Also, this request can not be undone.

execute_password_reset(reset_url, password)

Executes a password reset

This involves finding the password reset by its url, changing the password and then deleting the password reset.

Returns the Omscore.Auth.User with its updated password.

get_password_reset_by_url!(reset_url)

Finds a password reset object by its url.

You can not find it by the url directly as it's stored in hashed form.

Returns the Omscore.Auth.PasswordReset with the user preloaded.

get_user!(id)

Gets a user by its user ID.

Throws in case the id does not exist

Returns a Omscore.Auth.User.

get_user_by_email!(email)

Gets a user by its email address.

Casing of the email is ignored.

Returns a Omscore.Auth.User

get_user_by_member_id!(member_id)

Gets a user by its associated Omscore.Members.Member

You can either pass a string or an integer, if you pass a string it will attempt to parse it to int

Returns a Omscore.Auth.User

list_users()

Lists all Users in the system.

Returns an array of Omscore.Auth.Users with nothing preloaded.

login_user(username, password, device \\ "Unknown device")

Check user credentials and on success, generate a refresh and an access token.

The refresh token is valid for a long time (see config/config.exs) and can be used to obtain new access tokens The access token is valid for a short period of time and should be put into the x-auth-token header field to authenticate against oms services. Instead of the username, you can also use the email of the user. If you want, you can pass a device name to later recognize which refresh token belongs to which device, as it is stored in the refresh token claims

Returns {:ok, user, access_token, refresh_token} or {:error, _}

logout_token(refresh_token_id)

Invalidates a single refresh token.

Can be useful if a user has multiple token for different devices and only wants to log out on one of them

Returns {:ok, token} or {:error, nil} in case the refresh token was not found

logout_user(user)

Invalidates all refresh tokens of the user, completely logging him out.

Be careful that he can still posess valid access tokens, which can grant him access for one access token expiration phase.

Returns {:ok, deletes}, where deletes is the number of deleted refresh tokens

renew_token(refresh_token)

Get a new access token for a valid refresh token.

Returns {:ok, access} or {:error, _}

trigger_password_reset(email)

Triggers a password reset for a user, based on an email provided.

This will search the user (throw if nonexistent), remove all pending password resets for that user, create a new one and send a mail to the user with a url referencing that password reset request. The user then has to use Omscore.Auth.execute_password_reset/2 with the url he received in that mail to change his password.

Returns {:ok, password_reset}

update_user(user, attrs)

Updates an existing user.

You can not change the member id, superadmin field or active status through this. If you want to set the member id, use Omscore.Auth.update_user_member_id/2 If you want to set the superadmin status, use Omscore.Auth.update_user_superadmin/2 If you want to activate/deactivate a user, use Omscore.Auth.update_user_active/2

Returns {:ok, user} with the updated user.

update_user_active(user, active)

Updates whether a user is active or not.

Beware that existing access tokens are still valid, so it will take at most one access token expiration time until the user is actually deactivated.

Returns {:ok, user} with the updated user.

update_user_member_id(user, member_id)

Updates the associated member_id.

If you pass a binary, it will be parsed to integer.

Returns {:ok, user} with the updated user.

update_user_superadmin(user, superadmin)

Updates the superadmin status of an existing user.

Returns {:ok, user} with the updated user.