OMS Core Elixir v0.0.1 Omscore.Core

The Core context.

This module accumulates functions for bodies, circles and permissions and implements the permission system. For memberships, check out the Omscore.Members Module.

Functions

all_orphan_circles?(circles)

Checks if all circles in circles do not have a parent circle.

Useful to end a recursive permission check

Returns true if all circles are orphan or false if at least one is not.

apply_attribute_filters(data, filters)

Applies an attribute filter to any data

Filters should be a list of Omscore.Core.AttributeFilter. The data can be a map or a list, containing either string or atom keys In case fields from the filter are not present, they will be ommitted Filters are applied on nested items in case the field is of the form parent.child

Returns data but in filtered form.

change_body(body)

Creates a body changeset

change_circle(circle)

Creates a circle changeset.

change_permission(permission)

Creates a permission changeset

circles_have_same_body?(list)

Checks if all the circles are from the same body

Returns true or false

create_body(attrs \\ %{})

Creates a body

Ignores the shadow_circle, set it after creation with update_body/2.

Returns {:ok, body} or {:error, _}

create_circle(attrs \\ %{})

Creates a free circle

Returns {:ok, circle} or {:error, _}

create_circle(attrs, body)

Creates a bound circle

The circle will be bound to the passed body

Returns {:ok, circle} or {:error, _}

create_permission(attrs \\ %{})

Creates a permission

Returns {:ok, permission} or {:error, _}

delete_body(body)

Deletes a Body.

Cascades to body_memberships, bound circles and fee payments.

delete_circle(circle)

Deletes a Circle.

delete_permission(permission)

Deletes a permission, cascades to CirclePermissions.

find_array(type, input_data)

Converts an array that came from the user with maps that contain an id to a list of ecto models.

It loops through the input_data and tries to fetch every item from db by taking the "id" or :id field. You should specify which model to use by providing a type parameter. In case not all elements could be loaded, returns an error, even if some were successful.

Returns {:ok, [data]} or {:error, :not_found, message}

find_circles(input_data)

From an array of possibly circles, load those who are actually circles in the db

For more details, see find_array/2. This is useful for put_child_circles/2

find_permissions(input_data)

Finds all db permissions from a list of input permissions

Is a wrapper on find_array/2, see that method for more info.

Returns {:ok, [permissions]} or {:error, :not_found, _}

get_body!(id)

Gets a single body, circles preloaded

get_body_members(body)

This function is deprecated. Use `Omscore.Members.get_body_memberships/1` instead.

Returns all members of a body

get_child_circles(circles)

Returns all circles which are somewhere below in the inheritance tree.

You can either pass a circle or a list of circles.

Returns [circles]

get_circle(id)

Gets a single circle without throwing.

get_circle!(id)

Get a single circle

Preloads permission, child_circles and parent circle.

get_members_with_permission(id, params)

Returns all members which hold a certain permission in the system

This is scope-sensitive, so the local or global permission might yield other members. It takes into account circle permission inheritance, so it returns both directly and indirectly obtained permissions.

Returns [members]

get_permission(scope, action, object)

Gets a single permission by scope, object and action

Returns nil on non-existence.

get_permission!(id)

Gets a single permission

Throws on non-existence

get_permissions_recursive(circle)

Returns all permissions that are attached to this circle or any of its parent circles

This might be rather slow as it needs to perform several db requests.

Returns a list of permissions, possibly containing duplicates.

is_parent_recursive?(circle, parent_circle)

Checks if the parent circle actually is a parent of circle, somewhere in the inheritance tree.

Returns true or false

list_always_assigned_permissions()

Returns all always assigned permissions. You can also use list_permissions/1 and pass a filter for always_assigned

list_bodies(params \\ %{})

Returns all bodies in the system

You can pass filters, pagination or search params based on the format defined in OmscoreWeb.Helper

list_bound_circles(body, params \\ %{})

List all circles bound to a specific body

Supports filtering, search and pagination based on the syntax in OmscoreWeb.Helper.

Returns [circles]

list_bound_circles_(body, params)

list_bound_circles_with_permission(body, action, object)

Lists all bound circles which have a certain permission directly or indirectly attached.

This function is rather unperformant with big numbers of circles or deep inheritance structures. The permission is identified by action:object, scoping is ignored. This can be used to track which circle has a permission or to notify all members of that circle in case something happened to the object.

Returns [circles] where each circle holds his permissions in :all_permissions

list_circles(params \\ %{})

List all circles

Supports filtering, search and pagination based on the syntax in OmscoreWeb.Helper.

Returns [circles]

list_free_circles(params \\ %{})

List all unbound circles

Supports filtering, search and pagination based on the syntax in OmscoreWeb.Helper.

Returns [circles]

list_permissions(params \\ %{})

Returns all permissions in the system

You can pass filter, search or pagination parameters by the format defined in OmscoreWeb.Helper. Search is done automatically on object, action and scope

Returns a list.

put_child_circles(circle, child_circles)

Puts child circles for a circle

Removes the child-relation to existing children which are not in the list (but leaves the circles themselves intact, just nillifying the parent_circle_id) You do not have to sanitize data find_circles/1 if the data came from the user, as this is done internally child_circles can be a list of maps which somehow hold a :id or "id" field to an existing circle. If any of the circles was not found, it will fail and not change anything. If any of the circles was not an orphan circle, it will fail and not change anything.

Returns {:ok, circle} or {:error, code, msg}

put_circle_permissions(circle, permissions)

Directly assignes permissions to a circle

Expects a list of Omscore.Core.Permissions and assigns them as the new list of assigned permissions. It will unassign existing assigned permissions that are not in the list of permissions. You should preload the permission data with find_permissions if the data came from the user

Returns {:ok, circle} or {:error, _}

put_parent_circle(circle, parent_circle)

Puts the parent circle for a circle while maintaining joinable consistency

Maintaining joinable consistency means this method will fail in case the parent is non joinable and this circle is joinable

Returns {:ok, circle} or {:error, error-data}

reduce_permission_list(permission_list)

Reduces a permission list to remove duplicates

Expects a list of Omscore.Core.Permission and removes duplicates. If the same scope:action:object tuple exists in permission_list, the filters are intersected and the duplicate is removed. If a permission exists with different scopes but with the same action:object tuple, the one with the highest scope is returned after filters were intersected. Intersecting filters creates a theoretical loophole where a strongly filtered global permission intersected with an unfiltered local permission will result in an unfiltered global permission, so please don't check on scope anymore after reducing a permission list. Scope order is global-local-join_request.

If no duplicates existed, the list is left intact.

search_permission_list(permission_list, action, object)

Seaches a list of permissions for one specific action and object and returns it.

In case none was found, returns {:forbidden, "Permission xy required but not granted to you"} Ignores the scope of the permission. If you want a specific scope, use search_permission_list/4 Returns the first occurrence, use reduce_permission_list in advance in case you want the highest scoped one

Returns {:ok, permission} or {:forbidden, msg}

search_permission_list(permission_list, action, object, scope)

Searches a list of permissions for one specific scope:action:object and returns it

behaves like search_permission_list/3 but taking into account the scope aswell.

Returns {:ok, permission} or {:forbidden, msg}

update_body(body, attrs)

Updates a body

Returns {:ok, body} or {:error, _}

update_circle(circle, attrs)

Update a circle

You can only update name, description and joinable status with this method. If you want to update parent or child circles, use put_child_circles/2 or put_parent_circle/2. If you want to update assigned permissions, use put_circle_permissions/2.

Returns {:ok, circle} or {:error, _}.

update_permission(permission, attrs)

Updates a permission

Returns {:ok, permission} or {:error, _}