OMS Core Elixir v0.0.1 OmscoreWeb.Router.Helpers

Module with named helpers generated from OmscoreWeb.Router.

Functions

body_body_path(conn_or_endpoint, action, body_id)

body_body_path(conn_or_endpoint, action, body_id, params)

body_body_path(conn_or_endpoint, action, body_id, membership_id, params)

body_body_url(conn_or_endpoint, action, body_id)

body_body_url(conn_or_endpoint, action, body_id, params)

body_body_url(conn_or_endpoint, action, body_id, membership_id, params)

body_circle_path(conn_or_endpoint, action, body_id)

body_circle_path(conn_or_endpoint, action, body_id, params)

body_circle_url(conn_or_endpoint, action, body_id)

body_circle_url(conn_or_endpoint, action, body_id, params)

body_join_request_path(conn_or_endpoint, action, body_id)

body_join_request_path(conn_or_endpoint, action, body_id, params)

body_join_request_path(conn_or_endpoint, action, body_id, id, params)

body_join_request_url(conn_or_endpoint, action, body_id)

body_join_request_url(conn_or_endpoint, action, body_id, params)

body_join_request_url(conn_or_endpoint, action, body_id, id, params)

body_member_path(conn_or_endpoint, action, body_id)

body_member_path(conn_or_endpoint, action, body_id, params)

body_member_url(conn_or_endpoint, action, body_id)

body_member_url(conn_or_endpoint, action, body_id, params)

body_path(conn_or_endpoint, action)

body_path(conn_or_endpoint, action, params)

body_payment_path(conn_or_endpoint, action, body_id)

body_payment_path(conn_or_endpoint, action, body_id, params)

body_payment_path(conn_or_endpoint, action, body_id, payment_id, params)

body_payment_url(conn_or_endpoint, action, body_id)

body_payment_url(conn_or_endpoint, action, body_id, params)

body_payment_url(conn_or_endpoint, action, body_id, payment_id, params)

body_url(conn_or_endpoint, action)

body_url(conn_or_endpoint, action, params)

campaign_path(conn_or_endpoint, action)

campaign_path(conn_or_endpoint, action, params)

campaign_path(conn_or_endpoint, action, campaign_url, params)

campaign_path(conn_or_endpoint, action, campaign_url, submission_token, params)

campaign_url(conn_or_endpoint, action)

campaign_url(conn_or_endpoint, action, params)

campaign_url(conn_or_endpoint, action, campaign_url, params)

campaign_url(conn_or_endpoint, action, campaign_url, submission_token, params)

circle_path(conn_or_endpoint, action)

circle_path(conn_or_endpoint, action, params)

circle_path(conn_or_endpoint, action, circle_id, params)

circle_path(conn_or_endpoint, action, circle_id, membership_id, params)

circle_url(conn_or_endpoint, action)

circle_url(conn_or_endpoint, action, params)

circle_url(conn_or_endpoint, action, circle_id, params)

circle_url(conn_or_endpoint, action, circle_id, membership_id, params)

login_path(conn_or_endpoint, action)

login_path(conn_or_endpoint, action, params)

login_path(conn_or_endpoint, action, reset_url, params)

login_url(conn_or_endpoint, action)

login_url(conn_or_endpoint, action, params)

login_url(conn_or_endpoint, action, reset_url, params)

member_path(conn_or_endpoint, action)

member_path(conn_or_endpoint, action, params)

member_path(conn_or_endpoint, action, member_id, params)

member_url(conn_or_endpoint, action)

member_url(conn_or_endpoint, action, params)

member_url(conn_or_endpoint, action, member_id, params)

path(data, path)

Generates the path information including any necessary prefix.

payment_path(conn_or_endpoint, action)

payment_path(conn_or_endpoint, action, params)

payment_path(conn_or_endpoint, action, payment_id, params)

payment_url(conn_or_endpoint, action)

payment_url(conn_or_endpoint, action, params)

payment_url(conn_or_endpoint, action, payment_id, params)

permission_path(conn_or_endpoint, action)

permission_path(conn_or_endpoint, action, params)

permission_path(conn_or_endpoint, action, id, params)

permission_url(conn_or_endpoint, action)

permission_url(conn_or_endpoint, action, params)

permission_url(conn_or_endpoint, action, id, params)

static_path(conn, path)

Generates path to a static asset given its file path.

static_url(conn, path)

Generates url to a static asset given its file path.

url(data)

Generates the connection/endpoint base URL without any path information.